Update Thematic files
